Move user classes to data.models.usertypes
