+++ Disaster-Response-Platform/mobile/DisasterResponsePlatform/app/src/main/java/com/example/disasterresponseplatform/data/models/usertypes/RoleBasedUser.kt
@@ -1,4 +1,4 @@
-package com.example.disasterresponseplatform.models
+package com.example.disasterresponseplatform.data.models.usertypes
 
 class RoleBasedUser(username: String, email: String, phone: String, name: String, surname: String,
                     var proficiency: String): AuthenticatedUser(username, email, phone, name, surname) {
